chore: use RedisClient
